chore(*) define ValueWrapper as a commons usage object
