+++ cranberry-statement/src/main/java/io/github/ololx/cranberry/statement/internal/exception/NotBlankStatementException.java
@@ -16,6 +16,8 @@
  */
 package io.github.ololx.cranberry.statement.internal.exception;
 
+import io.github.ololx.cranberry.commons.wrapping.ValueWrapper;
+
 /**
  * The type Not blank statement exception.
  * Thrown to indicate that a not empty statement has failed.
