+++ cranberry-statement/src/main/java/io/github/ololx/cranberry/statement/internal/exception/StatementException.java
@@ -16,6 +16,8 @@
  */
 package io.github.ololx.cranberry.statement.internal.exception;
 
+import io.github.ololx.cranberry.commons.wrapping.ValueWrapper;
+
 import java.io.Serializable;
 import java.util.Arrays;
 import java.util.Locale;
@@ -32,249 +34,6 @@ import java.util.stream.Collectors;
  * @author Alexander A. Kropotin
  */
 public class StatementException extends IllegalArgumentException {
-
-    /**
-     * The type ValueWrapper.
-     * The wrapper of a value that is used in an statements exceptions.
-     * 
-     * This stores the value {@link #getValue()}, value type {@link #getType() runtime type}
-     * and value hash code {@link #toString()}.
-     * 
-     * This allows to get a value string representation {@link #getValueStringRepresentation()}
-     * 
-     * The {@link #toString()} method returns the string representation of the
-     * value along with its type and identity hash code.
-     * 
-     * project cranberry
-     * created 2020-04-16 12:45
-     * 
-     * @author Alexander A. Kropotin
-     */
-    static final class ValueWrapper implements Serializable {
-
-        /**
-         * Use serialVersionUID for interoperability
-         */
-        private static final long serialVersionUID = -309177L;
-
-        /**
-         * The wrapped value hash code; see {@link #toString()}.
-         * 
-         * @serial
-         */
-        private final int valueHashCode;
-
-        /**
-         * The wrapped value type; see {@link #toString()}, {@link #getType()}.
-         * 
-         * @serial
-         */
-        private final Class<?> type;
-
-        /**
-         * The wrapped value; see {@link #getValueStringRepresentation()},
-         * {@link #getValue()}, {@link #toString()}, {@link #getType()}.
-         * 
-         * @serial
-         */
-        private final Object value;
-
-        /**
-         * The factory method for the getting a new {@code ValueWrapper} instance
-         * for the supplied {@code value}.
-         * 
-         * This method will return a new {@code ValueWrapper} instance for a given value
-         * suitable for all {@code values} if the value is not an instance of a wrapper.
-         * Otherwise, if the supplied {@code value} is already a {@link ValueWrapper} instance,
-         * this method will return it as is.
-         * 
-         * @param value the value to wrap; may be {@code null}
-         * @return a wrapper for the supplied value; never {@code null}
-         */
-        public static ValueWrapper getInstance(Object value) {
-
-            if (value instanceof ValueWrapper)
-                return (ValueWrapper) value;
-
-            return new ValueWrapper(value);
-        }
-
-        /**
-         * Returns the string representation of a value for the supplied {@code value}
-         * via the supplied custom {@code nullStringRepresentation}.
-         * 
-         * If the supplied {@code value} is {@code null}, this method will return a
-         * custom representation of the value from the {@code nullStringRepresentation}.
-         * If the supplied {@code value} is not {@code null}, this method will return a string representation
-         * which is generated by invoking {@link String#valueOf(Object) String.valueOf(value)}:
-         * <ul>
-         *     <li>
-         *         for an each element of an array - if the value type is
-         *         an array type; see {@link Class#isArray()};
-         *     </li>
-         *     <li>
-         *         for the value - in another cases.
-         *     </li>
-         * </ul>
-         * 
-         * @param value the value to wrap; may be {@code null}
-         * @param nullStringRepresentation a custom representation of the value; is used by default
-         * when value is {@code null}.
-         * @return a string representation for the supplied value; never {@code null}
-         */
-        public static String getValueStringRepresentation(Object value, String nullStringRepresentation) {
-
-            if (value == null)
-                return String.valueOf(nullStringRepresentation);
-
-            if (value.getClass().isArray())
-                return String.format(
-                        "[%s]",
-                        Arrays.stream((Object[]) value)
-                                .map(element -> String.valueOf(element))
-                                .collect(Collectors.joining(", "))
-                );
-
-            return String.valueOf(value);
-        }
-
-        /**
-         * Constructs an {@code ValueWrapper} with the
-         * specified value, value type and value hash code.
-         * 
-         * @param value the value to wrap; may be {@code null}.
-         */
-        private ValueWrapper(Object value) {
-            this.valueHashCode = System.identityHashCode(value);
-            this.value = value;
-            this.type = value == null
-                    ? null
-                    : value.getClass();
-        }
-
-        /**
-         * Returns the origin value supplied to {@link #ValueWrapper(Object)} or
-         * {@code null} if the value is {@code null}.
-         * 
-         * @return an origin value with type {@code Object}.
-         */
-        public Object getValue() {
-            return this.value;
-        }
-
-        /**
-         * Returns the origin value type supplied to {@link #ValueWrapper(Object)} or
-         * {@code null} if the value is {@code null}.
-         * 
-         * @return a origin value type with type {@code Class<?>}.
-         */
-        public Class<?> getType() {
-            return this.type;
-        }
-
-        /**
-         * Returns the string representation of a wrapped value.
-         * This method returns wrapped value string representation by invoking
-         * {@link #getValueStringRepresentation(Object value, String nullStringRepresentation)
-         * ValueWrapper.getValueStringRepresentation(this.getValue(), null)}.
-         * 
-         * If the wrapped {@code value} is {@code null}, this method will return a
-         * string `{@code null}'.
-         * 
-         * @return a string representation for the wrapper value; never {@code null}
-         */
-        public String getValueStringRepresentation() {
-            return ValueWrapper.getValueStringRepresentation(this.getValue(), null);
-        }
-
-        /**
-         * Indicates whether the wrapped value is defined or not.
-         * 
-         * @return {@code true} if the wrapped value is not null;
-         * {@code false} otherwise.
-         */
-        public boolean isValueDefined() {
-            return this.getValue() != null;
-        }
-
-        /**
-         * Indicates whether the wrapped value is undefined or
-         * it is defined.
-         * 
-         * @return {@code true} if the wrapped value is null;
-         * {@code false} otherwise.
-         */
-        public boolean isValueUndefined() {
-            return !this.isValueDefined();
-        }
-
-        /**
-         * Indicates whether the wrapped value type is defined or not.
-         * 
-         * @return {@code true} if the wrapped value type is not null;
-         * {@code false} otherwise.
-         */
-        public boolean isTypeDefined() {
-            return this.getType() != null;
-        }
-
-        /**
-         * Indicates whether the wrapped value type is undefined
-         * or it is defined.
-         * 
-         * @return {@code true} if the wrapped value type is null;
-         * {@code false} otherwise.
-         */
-        public boolean isTypeUndefined() {
-            return !this.isTypeDefined();
-        }
-
-        /**
-         * Returns the string representation of a value along with its type and
-         * identity hash code.
-         * 
-         * The {@code toString} method for this class returns a string consisting of:
-         * <ul>
-         *     <li>
-         *         the string representation of the origin value - if the origin value is null;
-         *         
-         *         for instance:
-         *         <blockquote>
-         *             <pre>
-         *                 value.getValueStringRepresentation()
-         *             </pre>
-         *         </blockquote>
-         *     </li>
-         *     <li>
-         *         the name of the class of which the origin value object is an instance,
-         *         the at-sign character `{@code @}', the unsigned hexadecimal representation
-         *         of the hash code of the origin value object, the at-sign character `{@code =}'
-         *         and the string representation of the origin value - if the origin value is null.
-         *         
-         *         for instance:
-         *         <blockquote>
-         *             <pre>
-         *                 '[' + getClass().getName() + '@' + Integer.toHexString(value.hashCode()) + ']'
-         *                 + '=' + value.getValueStringRepresentation()
-         *             </pre>
-         *         </blockquote>
-         *     </li>
-         * </ul>
-         * 
-         * @return a string representation of the object.
-         */
-        @Override
-        public String toString() {
-            return this.isTypeDefined()
-                    ? String.format(
-                            "[%s@%s]=%s",
-                            this.getType().getName(),
-                            Integer.toHexString(this.valueHashCode),
-                            this.getValueStringRepresentation()
-                    )
-                    : this.getValueStringRepresentation();
-        }
-    }
 
     /**
      * Use serialVersionUID for interoperability
